Improved labels in the NodeSidePanel form
Using a label before an input with a `for` attribute
does not seem to work. `htmlFor` would be better, but
the React documentation recommends putting the input
*inside* the label. This makes clicking on a label
automatically focus the corresponding input.

Only the radio buttons keep the previous `<label htmlFor=...>` version, because it puts the radio
button *before* the label, which is better ergonomically.