Added colors to supporting/countering nodes
The color is (for now) only defined when creating
a node, and is not changed when the node is updated.

Perhaps it would be better to move these functions up
to the `App` component, so that we can use the same
colors when updating? (As the update function is defined
in `App`).

Also, this could help in the future, if we want to allow
settings to change the desired colors for supporting and
countering nodes: updating the settings should update
the `colors` dictionary *and* the nodes!