Added function to export the graph as Python code
The function simply transforms nodes into arguments (`Argument`) and edges
into attacks.
This allow exporting the (visual) graph into an executable reward function,
although it lacks some bits of code (only the AFDM creation is exported,
not the Reward class with the `compute` function and so on...

However, in this current form, the export is unidirectional: we cannot load
a graph from a Python file (at least not easily, without parsing the Python,
which can be quite difficult).
It could be improved by outputting a JSON-serialized form of the graph, which
would be far easier to parse back to a graph.