Refactored the node background color mechanism
Previously, the background color was set directly as the node's `style`.
However, that meant:
 * it was exported with the node's data when serialized (JSON, Python, ...);
 * we had to update the node's style when updating its data (redundancy of
   information...);
 * it was harder to allow for client-side customization of colors (e.g., a
   "settings" panel).

We can instead benefit from the `CustomNode` to directly render the
background color when rendering the node's content. We had to change
the node padding definition to render it correctly, because `CustomNode` is
actually the node's direct children, not the node itself.