Add random jitter when spawning nodes
Nodes are created at a default X/Y coords. Without adding jitter, they
would be placed on top of each other, and thus hide each other.
This was extra-confusing, because ReactFlow places the first node on top
of all others, which means that we would not see the creation of
subsequent nodes.

This commit adds a small random amount to both X/Y coords, so that
nodes should be at least slightly visible behind the others (although
there is a probability they spawn at the exact defaultX/defaultY coords).