Refactored method to update nodes
To update nodes, we previously used a lambda function defined in `App`
directly, so that it could rely on the `setNodes` method returned by
`useNodesState`. This meant we had to pass down this lambda to components
that want to update the nodes.

Instead, this commit defines a new function in the `nodes.js` file, similar
to the `createNewNode` and `duplicateNode`, which makes the code better
organized.
`App` can be free of this lambda, and components that want to update a node
simply import the function from `nodes.js`.