Added methods to save and load graphs in the browser storage
We use the local storage because it is persistent across sessions,
compared to the session storage which is cleaned when the session
is closed.
Currently, only "manual" saves are supported, i.e., saves created
explicitly by the user, but the module is designed to support other
types, e.g., "automatic" saves.