Added the "save / load" panel
This panel is used to save graphs in the browser storage (whereas
the import / export is used to interact with the user's file storage).

Note: we add `justifyContent: start` to align buttons within the button
group. Strangely, without any icon, they are aligned by default, but they
get centered when adding icons. Because they do not have the same size,
this centered alignment makes the group ugly.