Added the notistack library to easily create Snackbars (toasts)
Snackbars are an easy way to display a (short) message that should
not interrupt the user, and will automatically hide after a short
time. However, in MUI, a Snackbar must be created and controlled before
it can be shown (`open`, `setOpen`, `onClose`, etc.), which is quite
verbose, especially if we want to show several messages.

The notistack library simplifies this by allowing to create snackbars
simply by calling a method. In addition, it handles stacking multiple
snackbars.