Fixed ReactFlow size in new App layout
Because the AppBar has a higher zIndex, and the ReactFlow div internally
uses a height of "100%", we have 2 problems:

* the ReactFlow slightly slips below the AppBar, which is not important,
  because we can always scroll, de-zoom, etc.
* more importantly, it computes a wrong height, and thus overflows a bit,
  making the page scrollable, which we do NOT want (the whole content should
  fit as-is, the drawer itself is a scrollable portion of the fixed-size
  content).

This commit reduces the size of the box that encloses ReactFlow, with a
simple CSS `calc`, where the AppBar height is hardcoded (64px).
A better solution would be to set the AppBAr height in the theme, and then
get this value when computing the Box height.
A more ideal solution would be to simply put the ReactFlow *after* the
AppBar, and to take 100% of the "remaining" place...