Added documentation as a Quarto website
The tool might not be very intuitive to use, and needs a documentation
that, at the very least, explains the interface and how to use it, and
ideally explains a bit the underlying concepts (borrowed from AJAR;
without copy-pasting the research article).

However, writing such documentation in React would be very cumbersome...
The layout should be done "by hand" (even though components from MUI
could be used), and the very basic formatting (italic, bold, links, ...)
must also be done by hand.

Instead, Quarto allows creating a basic website very easily. The only
difficulty in this setup is that we must build the Quarto website alongside
the React index.html WebApp.

* Docs are written in a dedicated `docs/` folder to clearly separate the
  React WebApp from the Quarto Docs.
* Because we put the pages in an `articles/` folder, we must also create
  a redirection for the homepage (`index.html`). Quarto unfortunately does
  not have an easy way to say "I want my homepage to be path/to/doc.qmd"...