Added script to build both React WebApp and Quarto Docs
We previously only used `react-scripts build` to build the complete
webapp, thus it was sufficient to declare it as a one-liner in the
`package.json` configuration.

However, we now require to build both the React WebApp (through
`react-scripts`) AND the Quarto Docs (through `quarto render`),
and to pass the correct options to Quarto. Using a dedicated build script
is clearer.